(fileQueue) Extract localStorage preference readers to fileQueuePrefs
